+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Improve colour validation for icons
 - Fix theme colours for ruler colour
+- Remove trailing block comments when adding multiline todo to tree
 
 ## v0.0.206 - 2021-03-17